Add Import Progress Bar

+++ Source/Core/Editor/VisualEditor/Windows/GUI_Window_ImportProgressBar/GUI_Window_ImportProgressBar.h
@@ -32,11 +32,11 @@ class Window_ImportProgressBar {
 
     private:
 
-        long SubItemsImported_ = 0; /**<Set Number Of Items That Have Been Imported*/
-        long SubItemsToImport_ = 0; /**<Set Total Number Of Items Imported*/
+        int SubItemsImported_ = 0; /**<Set Number Of Items That Have Been Imported*/
+        int SubItemsToImport_ = 0; /**<Set Total Number Of Items Imported*/
         std::string CurrentAssetName_ = ""; /**<Set Current Asset Name*/
-        long CurrentAssetNumber_ = 0; /**<Set Current Number Of Asset Being Imported*/
-        long TotalAssetsToImport_ = 0; /**<Set Total Number Of Assets To Import*/
+        int CurrentAssetNumber_ = 0; /**<Set Current Number Of Asset Being Imported*/
+        int TotalAssetsToImport_ = 0; /**<Set Total Number Of Assets To Import*/
         std::mutex LockViewStats_; /**<Mutex To Prevent Values Being Changed During View/Write*/
 
         bool Enabled_ = true; /**<Is Popup Enabled*/
